Add calculations api and tests

+++ emissions-api/src/TechChallenge.Emissions.Api/Program.cs
@@ -5,6 +5,16 @@ using TechChallenge.Common.Exceptions;
 using TechChallenge.DataSimulator;
 using TechChallenge.Emissions.Api;
 using TechChallenge.Emissions.Api.Data;
+using Microsoft.Extensions.Logging;
+using System.Threading;
+using System.Collections.Generic;
+using System;
+using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Http.Features;
+using Microsoft.AspNetCore.Routing;
+using Microsoft.AspNetCore;
+using Microsoft.AspNetCore.Http.Extensions;
 
 var builder = WebApplication.CreateBuilder(args);
 
@@ -15,7 +25,7 @@ builder.Services.AddSwaggerGen();
 
 builder.Services.AddSingleton(TimeProvider.System);
 builder.Services.AddScoped<IEmissionsRepository, CalculationBasedEmissionsRepository>();
-var interval = (long)TimeSpan.FromMinutes(15).TotalSeconds;
+var interval = (int)TimeSpan.FromMinutes(15).TotalSeconds;
 builder.Services
     .AddSingleton<IValueCalculator<SeededContext, double>, RandomBasedDeterministicValueCalculator>()
     .AddSingleton<IPointsProvider, PointsProvider>(
